requirements.txt
# ===================================
# Bot de Trading Automatizado - Bybit
# Scalping de Futuros (Paper Mode)
# ===================================

# Core Trading Libraries
ccxt==4.2.25
pybit==5.6.2
websocket-client==1.7.0
requests==2.31.0

# Data Analysis & Technical Indicators
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
pandas-ta==0.3.14b0

# Async & WebSocket
websockets==12.0
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2

# UI & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Backtesting & Performance
backtrader==1.9.78.123
vectorbt==0.26.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
schedule==1.2.0

# Data Storage (optional)
sqlalchemy==2.0.25

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1